telegram_bot/requirements.txt
pytz
gino
aiogram
asyncio
asyncpg
pilmoji
python-dotenv
sqlalchemy==1.3